==> src/0-analyze_iotrace.ipynb <==
import pandas as pd

columns = ['ASU', 'SectorId', 'Size', 'Opcode', 'Timestamp']
df1 = pd.read_csv("../csv/raw/Financial1.csv", header=None, names=columns)
df2 = pd.read_csv("../csv/raw/Financial2.csv", header=None, names=columns)
print(f"Write operations of Financial1.csv: {len(df1['Opcode'] == 'w')}")
print(f"Write operations of Financial2.csv: {len(df2['Opcode'] == 'w')}")

# 데이터가 조금 더 많은 Financial1.csv 파일로 실험 결정
# iotrace_simulation 파일은 Financial1.csv 파일을 page 형식으로 바꾸고, model:simulation = 3:7로 나눈 파일이다.

columns = ['PageId', 'Size', 'Timestamp']
df_simulation = pd.read_csv("../csv/preprocessed/iotrace_simulation.csv", header=None, names=columns)
print(df_simulation.head())
print("I/O Trace size of simulation: ", len(df_simulation))
print("Unique PageIds of simulation: ", len(df_simulation['PageId'].unique()))
