==> src/2-get_partial_iotrace.ipynb <==
import pandas as pd

column = ['PageId', 'Size', 'Timestamp']
df = pd.read_csv('../csv/preprocessed/iotrace_simulation.csv', names=column)
df.head()

partial_df = df[:1000000]
partial_df.to_csv('../csv/preprocessed/iotrace_simulation_partial.csv', index=False)
