==> src/5-update_features_and_label.ipynb <==
import time
from collections import defaultdict

import numpy as np
import pandas as pd
import tensorflow as tf

import csv

def get_features():
    features = defaultdict(list)
    features_path = '../csv/preprocessed/features_timestamp_model.csv'
    str_to_float = lambda x: round(float(x), 10)
    with open(features_path, newline='', mode='r') as csvfile:
        reader = csv.reader(csvfile)
        next(reader) # 헤더 넘기기
        for page_id, frequency, timeinterval_avg, timeinterval_std, size_avg, last_timestamp in reader:
            page_id = int(page_id)
            frequency = int(frequency)
            timeinterval_avg = str_to_float(timeinterval_avg)
            timeinterval_std = str_to_float(timeinterval_std)
            size_avg = str_to_float(size_avg)
            last_timestamp = str_to_float(last_timestamp)
            features[page_id] = [frequency, timeinterval_avg, timeinterval_std, size_avg, last_timestamp]
    return features

def read_io_trace():
    io_trace = []
    io_trace_path = '../csv/preprocessed/iotrace_simulation.csv'
    str_to_float = lambda x: round(float(x), 10)
    with open(io_trace_path, newline='', mode='r') as csvfile:
        reader = csv.reader(csvfile)
        next(reader) # 첫 행 제거
        for logical_page_address, size, timestamp in reader:
            logical_page_address = int(logical_page_address)
            size = int(size)
            timestamp = str_to_float(timestamp)
            io_trace.append([logical_page_address, size, timestamp])

    return io_trace

def update_features(logical_page_address, size, timestamp):
    if logical_page_address not in features:
        features[logical_page_address] = [1, 3000, 3000, size, timestamp]
    else:
        frequency, timeinterval_avg, timeinterval_std, size_avg, last_timestamp = features[logical_page_address]
        timeinterval = timestamp - last_timestamp
        if frequency == 1:
            features[logical_page_address][1] = timeinterval
        else:
            features[logical_page_address][1] = float((timeinterval_avg * (frequency - 1) + timeinterval) / frequency)

        features[logical_page_address][0] = frequency + 1
        features[logical_page_address][3] = float((size_avg * frequency + size) / (frequency + 1))
        features[logical_page_address][4] = timestamp

def update_features_nofreq(logical_page_address, size, timestamp):
    if logical_page_address not in features:
        features[logical_page_address] = [1, 3000, 3000, size, timestamp]
    else:
        frequency, timeinterval_avg, timeinterval_std, size_avg, last_timestamp = features[logical_page_address]
        timeinterval = timestamp - last_timestamp
        if frequency == 1:
            features[logical_page_address][1] = timeinterval
        else:
            features[logical_page_address][1] = float((timeinterval_avg * (frequency - 1) + timeinterval) / frequency)

        features[logical_page_address][0] = frequency + 1
        features[logical_page_address][3] = float((size_avg * frequency + size) / (frequency + 1))
        features[logical_page_address][4] = timestamp

# features[page_id] = [frequency, timeinterval_avg, timeinterval_std, size_avg, last_timestamp]

features = get_features()
io_trace = read_io_trace()

decoded_dict = {0: 'Cold', 1: 'Hot', 2: 'Warm'}

def predict_with_h5(model_path) -> list:

    start = time.time()
    model = tf.keras.models.load_model(model_path)
    labeled_iotrace_simulation = []
    for idx, (logical_page_address, size, timestamp) in enumerate(io_trace):
        if idx % 1000 == 0:
            print(idx)

        update_features(logical_page_address, size, timestamp)

        input_features = np.array([features[logical_page_address][:-1]]) # last_timestamp 제외
        input_features = input_features.reshape((input_features.shape[0], 1, input_features.shape[1]))

        probabilities = model.predict(input_features, verbose=0) # verbose 0 : 진행상황 출력 x
        predict_label = np.argmax(probabilities, axis=1)[0]
        label = decoded_dict[predict_label]

        labeled_iotrace_simulation.append([logical_page_address, label])

    print("time elapsed: ", time.time() - start)

    return labeled_iotrace_simulation

def predict_with_pb(model_path) -> list:

    start = time.time()

    # 모델 불러오기
    loaded_model = tf.saved_model.load(model_path)

    # 모델의 서명 얻기
    infer = loaded_model.signatures["serving_default"]

    labeled_iotrace_simulation = []
    for idx, (logical_page_address, size, timestamp) in enumerate(io_trace):
        if idx % 10000 == 0:
            print(idx, time.time() - start)

        update_features(logical_page_address, size, timestamp)

        input_features = np.array([features[logical_page_address][1:-1]]) # last_timestamp 제외
        input_features = input_features.reshape((input_features.shape[0], 1, input_features.shape[1]))
        input_features = tf.cast(input_features, tf.float32)

        output = infer(tf.constant(input_features))
        probabilities = output[list(output.keys())[0]].numpy()
        label = np.argmax(probabilities, axis=1)[0]
        label = decoded_dict[label]

        labeled_iotrace_simulation.append([logical_page_address, label])

    print("time elapsed: ", time.time() - start)

    return labeled_iotrace_simulation

# predict_with_h5('./model/lstm.h5')
labeled_iotrace_simulation = predict_with_pb('../model/lstm_model/lstm_noFreq')

column = ['PageId', 'Label']
df = pd.DataFrame(labeled_iotrace_simulation, columns=column)
df.to_csv('../csv/preprocessed/labeled_iotrace_simulation.csv', index=False)

# 100만개에 812초 : 15분 .

df = pd.read_csv('../csv/preprocessed/labeled_iotrace_simulation.csv')
df.head()

print(len(df[df['Label'] == 'Cold']))
print(len(df[df['Label'] == 'Hot']))
print(len(df[df['Label'] == 'Warm']))
